// source/client/routes/static/odata/routes/static/Transport/Transport/Transport.ts
export type Transport = {
	readonly acceptanceStatus: 0;
	readonly archivalPortal: false;
	readonly attachmentsExist: true;
	readonly cancelled: false;
	readonly cargoAmount: ``;
	readonly carrier: string;
	readonly carrierAcceptation: null;
	readonly carrierAcceptationDate: null;
	readonly carrierAcceptationUser: null;
	readonly carrierAcceptationUserId: null;
	readonly carrierCharges: 0;
	readonly carrierChargesCurrency: string;
	readonly carrierChargesWitoutDiscount: 0;
	readonly carrierChargesWitoutDiscountCurrency: string;
	readonly carrierDaysOfDefermentOfPayment: 0;
	readonly carrierFAAccount: ``;
	readonly carrierFullName: string;
	readonly carrierId: number;
	readonly carrierReferenceNumber: ``;
	readonly carrierRefNo: ``;
	readonly commentsForCarrier: ``;
	readonly commentsFromCarrier: null;
	readonly companyId: number;
	readonly costInvoiceNumber: ``;
	readonly costInvoiceStatus: ``;
	readonly customer: string;
	readonly customerLoadedKm: number;
	readonly description: ``;
	readonly discount: 0;
	readonly driver1: string;
	readonly driver1BusinessPhone: ``;
	readonly driver1Id: number;
	readonly driver1IdentityCard: ` `;
	readonly driver2: ``;
	readonly driver2BusinessPhone: null;
	readonly driver2Id: number;
	readonly driver2IdentityCard: ` `;
	readonly eInvoiceAcceptance: false;
	readonly externalNumber: ``;
	readonly finishDate: Date;
	readonly finishDateOrder: 0;
	readonly finishDateRange: `00:00`;
	readonly finishRealDate: null;
	readonly gmp: false;
	/**
	 * Id
	 */
	readonly id: number;
	readonly idKind: number;
	readonly importOrigin: 0;
	readonly invoicedCosts: false;
	readonly invoiceState: ``;
	readonly kmTotal: 0;
	readonly kmWithCargo: 0;
	readonly kmWithoutCargo: 0;
	readonly loadingTime: `00:00`;
	readonly ltl: false;
	readonly modificationDate: null;
	readonly number: string;
	readonly numericalNumber: number;
	readonly numericalTransportNumber: number;
	readonly package: ``;
	readonly portalAccess: false;
	readonly portalAttachmentsExist: false;
	readonly requirements: ``;
	readonly route: string;
	readonly semiTrailer: ``;
	readonly semiTrailerId: 0;
	readonly startDate: Date;
	readonly startDateOrder: 0;
	readonly startDateRange: `00:00`;
	readonly startRealDate: null;
	readonly state: 1;
	readonly statusKindDescription: null;
	readonly statusKindId: null;
	readonly tapa: false;
	readonly toll: 0;
	readonly tollCurrency: string;
	readonly tollType: false;
	readonly trailer: ``;
	readonly trailerId: 0;
	readonly transportAuthorId: number;
	readonly transportNumber: string;
	readonly truck: string;
	readonly truckId: number;
	readonly unitAmount: number;
	readonly unitAmountType: null;
	readonly unitPrice: 0;
	readonly unitPriceCurrency: string;
	readonly unloadingTime: `00:00`;
	readonly weightInKg: 0;
};